servers/s3/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pytz
python-dateutil
boto3
pandas
pyarrow